Fixed cargo hold issues, Cargo hold is now on player
