+++ app/src/main/java/com/github/esebs/cs2340project/spacetrader/viewmodels/TradingViewModel.java
@@ -35,7 +35,7 @@ public class TradingViewModel {
      * @return the quantity of the resource owned by the Player
      */
     public int getSellQuantity(Resource resource) {
-        return vehicle.getCargoHold()[resource.ordinal()];
+        return player.getCargoHold()[resource.ordinal()];
     }
 
     /**
@@ -80,7 +80,7 @@ public class TradingViewModel {
         int credits = player.getCredits();
         maxBuyQuantity = Math.min(credits / price, maxBuyQuantity);
 
-        int remainingCargoSpace = vehicle.calculateRemainingCargoSpace();
+        int remainingCargoSpace = player.calculateRemainingCargoSpace();
 
         return Math.min(maxBuyQuantity, remainingCargoSpace);
     }
@@ -95,7 +95,7 @@ public class TradingViewModel {
      * @return the max sell quantity of the given resource
      */
     public int calculateMaxSellQuantity(Resource resource) {
-        int[] cargoHold = vehicle.getCargoHold();
+        int[] cargoHold = player.getCargoHold();
         return cargoHold[resource.ordinal()];
     }
 
@@ -108,9 +108,9 @@ public class TradingViewModel {
      * @param numToBuy how many to buy
      */
     public void buyResources(Resource resource, int numToBuy) {
-        int[] cargoHold = vehicle.getCargoHold();
+        int[] cargoHold = player.getCargoHold();
         cargoHold[resource.ordinal()] += numToBuy;
-        vehicle.setCargoHold(cargoHold);
+        player.setCargoHold(cargoHold);
 
         int credits = player.getCredits();
         int costPerUnit = currentRoom.getBuyFromRoomPrices()[resource.ordinal()];
@@ -134,9 +134,9 @@ public class TradingViewModel {
      * @param numToSell how many to sell
      */
     public void sellResources(Resource resource, int numToSell) {
-        int[] cargoHold = vehicle.getCargoHold();
+        int[] cargoHold = player.getCargoHold();
         cargoHold[resource.ordinal()] -= numToSell;
-        vehicle.setCargoHold(cargoHold);
+        player.setCargoHold(cargoHold);
 
         int credits = player.getCredits();
         int payPerUnit = currentRoom.getSellToRoomPrices()[resource.ordinal()];
